Modificacion en DAO y DAOJPA al tipo de parametro que recibe el metodo byid, cambio de long a int

+++ src/modelo/Bag.java
@@ -35,7 +35,7 @@ public class Bag
 	}
 	
 	//Retorno de una cartera por ID
-	public void getPaymentID(long bagID) throws Exception 
+	public void getPaymentID(int bagID) throws Exception 
 	{
 		_bag.byId(bagID);
 	}
